move vertex functions to general
