+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3vertex.cpp
@@ -1,4 +1,4 @@
-/** Copyright (C) 2015 Harijs Grinbergs
+/** Copyright (C) 2018 Robert B. Colton
 ***
 *** This file is a part of the ENIGMA Development Environment.
 ***
@@ -15,39 +15,9 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
-#include "../General/OpenGLHeaders.h"
-#include <vector>
-#include <map>
-#include <utility>
-using std::vector;
-using std::pair;
-
-#include "../General/GSvertex.h"
-#include "../General/GStextures.h"
-
-namespace enigma {
-  vector<VertexFormat*> vertexFormats(0);    
-}
+#include "Graphics_Systems/General/OpenGLHeaders.h"
+#include "Graphics_Systems/General/GSvertex.h"
 
 namespace enigma_user {
-
-  unsigned vertex_format_create(){
-    unsigned id = enigma::vertexFormats.size();
-    enigma::vertexFormats.push_back(new enigma::VertexFormat());
-    return id;
-  }
-
-  void vertex_format_destroy(int id){
-    delete enigma::vertexFormats[id];
-    enigma::vertexFormats[id] = nullptr;
-  }
-
-  bool vertex_format_exists(int id){
-    return (id >= 0 && (unsigned)id < enigma::vertexFormats.size() &&  enigma::vertexFormats[id] != nullptr);
-  }
-
-  void vertex_format_add(int id, int type, int attribute) {
-    enigma::vertexFormats[id]->AddAttribute(type, attribute);
-  }
 
 }
